pytorch_requirements.txt
# Core ML and RL dependencies - PyTorch version for HPRC compatibility
# Python 3.8+ with modern PyTorch stack

# Core computation
numpy>=1.24.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# Modern Gym and MuJoCo environments  
gymnasium>=0.29.0
mujoco>=2.3.0

# Scientific computing
scipy>=1.10.0

# Command line utilities
click>=8.1.0
tqdm>=4.65.0

# Visualization and plotting  
matplotlib>=3.6.0
seaborn>=0.12.0

# Computer vision and image processing
opencv-python>=4.8.0
pillow>=9.5.0
imageio>=2.31.0

# Logging and monitoring
tensorboard>=2.14.0

# Data analysis
pandas>=2.0.0

# Utilities
joblib>=1.3.0
dill>=0.3.7
cloudpickle>=2.2.1

# Progress bars
progressbar2>=4.2.0

# Optional gaming/simulation support
pygame>=2.1.0
ale-py>=0.8.1

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Note: All versions are compatible with Python 3.8+ and modern HPRC environments
# Removed: mpi4py, tensorflow, protobuf, baselines (replaced with custom implementations)
# Added: torch, gymnasium, tensorboard for modern ML stack